changed some ints to long ints

+++ modular/source/3DXY/3DXYlattice.h
@@ -5,7 +5,7 @@
 struct Lattice3DXY{
 	long double *** theLattice;
 	long double L,Nspins,Neqsweeps,NTotSweeps;
-	int Neqclusts,NTotClusts;	
+	long int Neqclusts,NTotClusts;	
 	bool coldstart,warmedUp;
 
 	long double energy,xmag,ymag,sinx,siny,sinz;
